requirements.txt
amqp==5.0.9
azure-core==1.21.1
azure-storage-blob==12.9.0
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==36.0.1
greenlet==1.1.2
idna==3.3
isodate==0.6.1
kombu==5.2.3
mccabe==0.6.1
msrest==0.6.21
mysql-connector-python==8.0.27
oauthlib==3.1.1
prompt-toolkit==3.0.24
protobuf==3.19.1
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.0
pyflakes==2.4.0
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.0
requests-oauthlib==1.3.0
six==1.16.0
SQLAlchemy==1.4.29
toml==0.10.2
typing_extensions==4.0.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
